Move test utilities from `testutil-server` to tests of `server`
