+++ pom.xml
@@ -12,7 +12,7 @@ all modules and does not describe the project structure per-subproject.
 
 <groupId>io.spine</groupId>
 <artifactId>spine-core-java</artifactId>
-<version>1.0.1</version>
+<version>1.0.2-SNAPSHOT</version>
 
 <inceptionYear>2015</inceptionYear>
