+++ server/src/test/java/io/spine/server/aggregate/given/dispatch/AggregatePartBuilder.java
@@ -18,15 +18,15 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.testing.server.aggregate;
+package io.spine.server.aggregate.given.dispatch;
 
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import com.google.protobuf.Message;
 import io.spine.server.aggregate.AggregatePart;
 import io.spine.server.aggregate.AggregateRoot;
 import io.spine.server.aggregate.model.AggregatePartClass;
+import io.spine.server.entity.EntityBuilder;
 import io.spine.server.entity.model.EntityClass;
-import io.spine.testing.server.entity.EntityBuilder;
 
 import java.lang.reflect.Constructor;
