+++ server/src/test/java/io/spine/server/procman/given/dispatch/package-info.java
@@ -19,12 +19,12 @@
  */
 
 /**
- * This package provides test utilities and base classes for the unit tests of
- * {@linkplain io.spine.server.procman.ProcessManager process managers}.
+ * Contains test utilities for creating instances of Process Managers and dispatching signals.
  */
+
 @CheckReturnValue
 @ParametersAreNonnullByDefault
-package io.spine.testing.server.procman;
+package io.spine.server.procman.given.dispatch;
 
 import com.google.errorprone.annotations.CheckReturnValue;
