+++ server/src/test/java/io/spine/server/projection/ProjectionTest.java
@@ -31,6 +31,7 @@ import io.spine.core.Versions;
 import io.spine.server.BoundedContext;
 import io.spine.server.BoundedContextBuilder;
 import io.spine.server.DefaultRepository;
+import io.spine.server.entity.given.Given;
 import io.spine.server.entity.storage.EntityColumn;
 import io.spine.server.entity.storage.EntityColumnCache;
 import io.spine.server.model.DuplicateHandlerMethodError;
@@ -60,7 +61,6 @@ import io.spine.test.projection.event.Int32Imported;
 import io.spine.test.projection.event.StringImported;
 import io.spine.testing.TestValues;
 import io.spine.testing.server.TestEventFactory;
-import io.spine.testing.server.entity.given.Given;
 import io.spine.type.TypeUrl;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -77,6 +77,7 @@ import static io.spine.protobuf.AnyPacker.pack;
 import static io.spine.server.projection.given.ProjectionTestEnv.FilteringProjection.SET_A;
 import static io.spine.server.projection.given.ProjectionTestEnv.FilteringProjection.SET_B;
 import static io.spine.server.projection.given.ProjectionTestEnv.NoDefaultOptionProjection.ACCEPTED_VALUE;
+import static io.spine.server.projection.given.dispatch.ProjectionEventDispatcher.dispatch;
 import static io.spine.server.projection.model.ProjectionClass.asProjectionClass;
 import static io.spine.server.storage.LifecycleFlagField.archived;
 import static io.spine.server.storage.LifecycleFlagField.deleted;
@@ -85,7 +86,6 @@ import static io.spine.server.type.given.GivenEvent.withMessage;
 import static io.spine.test.projection.Project.Status.STARTED;
 import static io.spine.testing.TestValues.random;
 import static io.spine.testing.TestValues.randomString;
-import static io.spine.testing.server.projection.ProjectionEventDispatcher.dispatch;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
