+++ server/src/test/java/io/spine/server/projection/given/dispatch/ProjectionEventDispatcher.java
@@ -17,7 +17,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.testing.server.projection;
+package io.spine.server.projection.given.dispatch;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.protobuf.Message;
