+++ server/src/test/java/io/spine/server/stand/MultitenantStandTest.java
@@ -38,8 +38,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import static io.spine.server.entity.given.Given.aggregateOfClass;
 import static io.spine.server.stand.given.StandTestEnv.newStand;
-import static io.spine.testing.server.entity.given.Given.aggregateOfClass;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
