+++ server/src/test/java/io/spine/server/stand/StandTest.java
@@ -105,14 +105,14 @@ import static io.spine.client.TopicValidationError.UNSUPPORTED_TOPIC_TARGET;
 import static io.spine.grpc.StreamObservers.memoizingObserver;
 import static io.spine.grpc.StreamObservers.noOpObserver;
 import static io.spine.protobuf.AnyPacker.unpack;
+import static io.spine.server.entity.given.Given.aggregateOfClass;
+import static io.spine.server.entity.given.Given.projectionOfClass;
 import static io.spine.server.stand.given.Given.StandTestProjection;
 import static io.spine.server.stand.given.StandTestEnv.newStand;
 import static io.spine.test.projection.Project.Status.CANCELLED;
 import static io.spine.test.projection.Project.Status.STARTED;
 import static io.spine.test.projection.Project.Status.UNDEFINED;
 import static io.spine.testing.Tests.assertMatchesMask;
-import static io.spine.testing.server.entity.given.Given.aggregateOfClass;
-import static io.spine.testing.server.entity.given.Given.projectionOfClass;
 import static io.spine.validate.Validate.isNotDefault;
 import static java.lang.String.valueOf;
 import static java.util.Arrays.asList;
